React skeleton with routing.
