+++ src/SpellsReference/package.json
@@ -21,7 +21,8 @@
   },
   "dependencies": {
     "react": "^16.10.2",
-    "react-dom": "^16.10.2"
+    "react-dom": "^16.10.2",
+    "react-router-dom": "^5.1.2"
   },
   "-vs-binding": {
     "BeforeBuild": [
